逆向工程生成从数据库取数据

+++ manong/manong_parent/manong_manager/manong_manager_mapper/src/main/java/com/manong/mapper/ProductCategoryMapper.java
@@ -11,7 +11,7 @@ public interface ProductCategoryMapper {
 
     int deleteByExample(ProductCategoryExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Short id);
 
     int insert(ProductCategory record);
 
@@ -19,7 +19,7 @@ public interface ProductCategoryMapper {
 
     List<ProductCategory> selectByExample(ProductCategoryExample example);
 
-    ProductCategory selectByPrimaryKey(Integer id);
+    ProductCategory selectByPrimaryKey(Short id);
 
     int updateByExampleSelective(@Param("record") ProductCategory record, @Param("example") ProductCategoryExample example);
